NexoraSmartCatalog/requirements.txt
reportlab==4.1.0
qrcode==7.4.2
Pillow==10.2.0